[SW-2008] Fix compile after removing code for external backend from H2O
